added verify as middleware
